feat: adiciona novas funcionalidade de enviar notificação ao cliente após venda
