+++ ExercicioDoisInterface/Empresa.cs
@@ -20,5 +20,13 @@ namespace ExercicioDoisInterface
             CNPJ = cNPJ ?? throw new ArgumentNullException(nameof(cNPJ));
             Setor = setor ?? throw new ArgumentNullException(nameof(setor));
         }
+
+        public void ImprimeInfo()
+        {
+            Console.WriteLine($"Nome Fantasia: {NomeFantasia}");
+            Console.WriteLine($"Razão Social: {RazaoSocial}");
+            Console.WriteLine($"CNPJ: {CNPJ}");
+            Console.WriteLine($"Setor: {Setor}");
+        }
     }
 }
