+++ ExercicioDoisInterface/Varejo.cs
@@ -8,21 +8,16 @@ namespace ExercicioDoisInterface
 {
     internal class Varejo : Empresa, IEmpresa
     {
-        public Varejo(string nomeFantasia, string razaoSocial, string cNPJ, string setor) : base(nomeFantasia, razaoSocial, cNPJ, setor)
+        public EnviarNotificacaoWhatsapp Notificacao { get; }
+        public Varejo(string nomeFantasia, string razaoSocial, string cNPJ, string setor, EnviarNotificacaoWhatsapp notificacao) : base(nomeFantasia, razaoSocial, cNPJ, setor)
         {
-        }
-
-        public void ImprimeInfo()
-        {
-            Console.WriteLine($"Nome Fantasia: {NomeFantasia}");
-            Console.WriteLine($"Razão Social: {RazaoSocial}");
-            Console.WriteLine($"CNPJ: {CNPJ}");
-            Console.WriteLine($"Setor: {Setor}");
+            Notificacao = notificacao;
         }
 
         public void RealizarVenda()
         {
             Console.WriteLine("Venda de PRODUTOS realizada!");
+            Notificacao.EnviarMensagemCliente();
         }
     }
 }
